Split expert insight tests to own file to avoid shared db tear down issue
